+++ spec/fixtures/Sample Project/ContainsSubproject/ContainsSubproject/AppDelegate.swift
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.swift
+//  ContainsSubproject
+//
+//  Created by Marius Rackwitz on 14/09/14.
+//  Copyright (c) 2014 Marius Rackwitz. All rights reserved.
+//
+
+import UIKit
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    var window: UIWindow?
+
+}
+