+++ spec/project/object/container_item_proxy_spec.rb
@@ -28,5 +28,44 @@ module ProjectSpecs
       @proxy.remote_info.should == 'Pods'
     end
 
+    describe '#remote?' do
+      it 'returns false if the container is for the current project' do
+        @proxy.container_portal = @project.root_object.uuid
+        @proxy.remote?.should.be.false
+      end
+
+      it 'returns true if the container is for a subproject' do
+        path = fixture_path('Sample Project/ReferencedProject/ReferencedProject.xcodeproj')
+        subproject = Xcodeproj::Project.open(path)
+        @project.main_group.new_file(path)
+        @proxy.container_portal = subproject.root_object.uuid
+
+        @proxy.remote?.should.be.true
+      end
+    end
+
+    describe '#proxied_object' do
+      before do
+        subproject_path = fixture_path('Sample Project/ReferencedProject/ReferencedProject.xcodeproj')
+        @subproject = Xcodeproj::Project.open(subproject_path)
+
+        path = fixture_path('Sample Project/ContainsSubproject/ContainsSubproject.xcodeproj')
+        @project = Xcodeproj::Project.open(path)
+      end
+
+      it 'returns the proxied object if it is contained in the current project' do
+        target = @project.targets.find { |t| t.name == 'ContainsSubprojectTests' }
+        @proxy = target.dependencies.first.target_proxy
+        proxied_object = @proxy.proxied_object
+        proxied_object.should == @project.targets.first
+      end
+
+      it 'returns the proxied object if it is contained in a subproject' do
+        @proxy = @project.targets.first.dependencies.first.target_proxy
+        proxied_object = @proxy.proxied_object
+        proxied_object.should == @subproject.targets.first
+      end
+    end
+
   end
 end